# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * GROFO API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiErrorBody {
  error: string;
}

export type VideoCategory = typeof VideoCategory[keyof typeof VideoCategory];


export const VideoCategory = {
  NEET: 'NEET',
  JEE: 'JEE',
  CLAT: 'CLAT',
  UPSC: 'UPSC',
  Languages: 'Languages',
} as const;

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Profile {
  displayName: string;
  /** @nullable */
  examName?: string | null;
  /** @nullable */
  examDate?: string | null;
  avatarColor: string;
  createdAt: string;
}

export interface ProfileUpdate {
  /** @minLength 1 */
  displayName?: string;
  /** @nullable */
  examName?: string | null;
  /** @nullable */
  examDate?: string | null;
  avatarColor?: string;
}

export interface Video {
  id: number;
  youtubeId: string;
  title: string;
  channel: string;
  category: VideoCategory;
  durationSeconds: number;
  thumbnailUrl: string;
}

export interface ContinueWatchingItem {
  videoId: number;
  video: Video;
  progressSeconds: number;
  lastWatchedAt: string;
}

export interface DashboardSummary {
  displayName: string;
  todayStudyMinutes: number;
  currentStreak: number;
  longestStreak: number;
  /** @nullable */
  examName: string | null;
  /** @nullable */
  examDate: string | null;
  /** @nullable */
  examDaysLeft: number | null;
  continueLearning: ContinueWatchingItem | null;
  todayTasksTotal: number;
  todayTasksCompleted: number;
  weeklyProgressPercent: number;
}

export interface Task {
  id: number;
  title: string;
  date: string;
  completed: boolean;
  priority: TaskPriority;
  createdAt: string;
}

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  date: string;
  priority?: TaskPriority;
}

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  date?: string;
  completed?: boolean;
  priority?: TaskPriority;
}

export type StudySessionSource = typeof StudySessionSource[keyof typeof StudySessionSource];


export const StudySessionSource = {
  manual: 'manual',
  pomodoro: 'pomodoro',
} as const;

export interface StudySession {
  id: number;
  durationMinutes: number;
  source: StudySessionSource;
  date: string;
  createdAt: string;
}

export interface StudySessionInput {
  /** @minimum 1 */
  durationMinutes: number;
  source: StudySessionSource;
}

export interface DailyStudyMinutes {
  date: string;
  minutes: number;
}

export interface ContinueWatchingInput {
  /** @minimum 0 */
  progressSeconds: number;
}

export interface BookmarkInput {
  videoId: number;
}

export interface Playlist {
  id: number;
  name: string;
  itemCount: number;
  createdAt: string;
}

export interface PlaylistInput {
  /** @minLength 1 */
  name: string;
}

export interface PlaylistItemInput {
  videoId: number;
}

export interface PlaylistDetail {
  id: number;
  name: string;
  createdAt: string;
  videos: Video[];
}

export interface NoteFolder {
  id: number;
  name: string;
  createdAt: string;
}

export interface NoteFolderInput {
  /** @minLength 1 */
  name: string;
}

export interface Note {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  folderId: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface NoteInput {
  /** @minLength 1 */
  title: string;
  content?: string;
  /** @nullable */
  folderId?: number | null;
}

export interface NoteUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  /** @nullable */
  folderId?: number | null;
}

export interface Material {
  id: number;
  title: string;
  category: VideoCategory;
  sourceUrl: string;
  /** @nullable */
  pageCount: number | null;
}

export interface MaterialBookmark {
  id: number;
  materialId: number;
  pageNumber: number;
  /** @nullable */
  note: string | null;
  createdAt: string;
}

export interface MaterialBookmarkInput {
  /** @minimum 1 */
  pageNumber: number;
  /** @nullable */
  note?: string | null;
}

export interface MaterialHighlight {
  id: number;
  materialId: number;
  pageNumber: number;
  text: string;
  color: string;
  createdAt: string;
}

export interface MaterialHighlightInput {
  /** @minimum 1 */
  pageNumber: number;
  /** @minLength 1 */
  text: string;
  color: string;
}

export type WorkspaceLayoutType = typeof WorkspaceLayoutType[keyof typeof WorkspaceLayoutType];


export const WorkspaceLayoutType = {
  'video-notes': 'video-notes',
  'video-pdf': 'video-pdf',
  'pdf-notes': 'pdf-notes',
  'ai-pdf': 'ai-pdf',
  'ai-notes': 'ai-notes',
} as const;

export interface WorkspaceLayout {
  contextKey: string;
  layoutType: WorkspaceLayoutType;
  splitRatio: number;
  /** @nullable */
  noteId: number | null;
  updatedAt: string;
}

export interface WorkspaceLayoutInput {
  layoutType: WorkspaceLayoutType;
  /**
     * @minimum 0
     * @maximum 100
     */
  splitRatio: number;
  /** @nullable */
  noteId?: number | null;
}

export type AiSessionMode = typeof AiSessionMode[keyof typeof AiSessionMode];


export const AiSessionMode = {
  explain: 'explain',
  doubt: 'doubt',
  summarize: 'summarize',
  quiz: 'quiz',
} as const;

export interface AiSession {
  id: number;
  title: string;
  mode: AiSessionMode;
  createdAt: string;
}

export interface AiSessionInput {
  /** @minLength 1 */
  title: string;
  mode: AiSessionMode;
}

export type AiMessageRole = typeof AiMessageRole[keyof typeof AiMessageRole];


export const AiMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface AiMessage {
  id: number;
  sessionId: number;
  role: AiMessageRole;
  content: string;
  createdAt: string;
}

export interface AiMessageInput {
  /** @minLength 1 */
  content: string;
}

export interface Achievement {
  id: string;
  title: string;
  description: string;
  achieved: boolean;
  /** @nullable */
  achievedAt: string | null;
}

export type ListTasksParams = {
startDate?: string;
endDate?: string;
};

export type ListVideosParams = {
category?: VideoCategory;
search?: string;
};

export type ListNotesParams = {
folderId?: number;
search?: string;
};

export type ListMaterialsParams = {
category?: VideoCategory;
search?: string;
};