# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GROFO API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /continue-watching/{videoId}:
    parameters:
      - name: videoId
        in: path
        required: true
        schema:
          type: integer
    put:
      operationId: upsertContinueWatching
      tags: [continue-watching]
      summary: Save/update watch progress for a video
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContinueWatchingInput"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContinueWatchingItem"
  /bookmarks:
    get:
      operationId: listBookmarks
      tags: [bookmarks]
      summary: List bookmarked videos
      responses:
        "200":
          description: Bookmarked videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
    post:
      operationId: createBookmark
      tags: [bookmarks]
      summary: Bookmark a video
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkInput"
      responses:
        "201":
          description: Bookmarked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
  /bookmarks/{videoId}:
    parameters:
      - name: videoId
        in: path
        required: true
        schema:
          type: integer
    delete:
      operationId: deleteBookmark
      tags: [bookmarks]
      summary: Remove a video bookmark
      responses:
        "204":
          description: Deleted
...
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: Browse the PDF study material catalog
      parameters:
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/VideoCategory"
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
  /materials/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getMaterial
      tags: [materials]
      summary: Get a single material
      responses:
        "200":
          description: Material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /materials/{id}/bookmarks:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: listMaterialBookmarks
      tags: [materials]
      summary: List bookmarks for a PDF material
      responses:
        "200":
          description: Material bookmarks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaterialBookmark"
    post:
      operationId: createMaterialBookmark
      tags: [materials]
      summary: Bookmark a page in a PDF material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialBookmarkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialBookmark"
...
  /materials/{id}/highlights:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: listMaterialHighlights
      tags: [materials]
      summary: List highlights for a PDF material
      responses:
        "200":
          description: Material highlights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaterialHighlight"
    post:
      operationId: createMaterialHighlight
      tags: [materials]
      summary: Highlight text on a PDF page
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialHighlightInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialHighlight"
...
  /ai/sessions:
    get:
      operationId: listAiSessions
      tags: [ai]
      summary: List AI Sensei chat sessions
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiSession"
    post:
      operationId: createAiSession
      tags: [ai]
      summary: Start a new AI Sensei chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiSessionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSession"
  /ai/sessions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    delete:
      operationId: deleteAiSession
      tags: [ai]
      summary: Delete an AI Sensei chat session
      responses:
        "204":
          description: Deleted
  /ai/sessions/{id}/messages:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: listAiMessages
      tags: [ai]
      summary: List messages in an AI Sensei session
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiMessage"
    post:
      operationId: sendAiMessage
      tags: [ai]
      summary: Send a message and get the AI Sensei's reply
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiMessageInput"
      responses:
        "201":
          description: Assistant reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiMessage"
        "503":
          description: AI Sensei is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorBody"
...
components:
  schemas:
...
    ApiErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
    VideoCategory:
      type: string
      enum: [NEET, JEE, CLAT, UPSC, Languages]
...
    Video:
      type: object
      required:
        - id
        - youtubeId
        - title
        - channel
        - category
        - durationSeconds
        - thumbnailUrl
      properties:
        id:
          type: integer
        youtubeId:
          type: string
        title:
          type: string
        channel:
          type: string
        category:
          $ref: "#/components/schemas/VideoCategory"
        durationSeconds:
          type: integer
        thumbnailUrl:
          type: string
    ContinueWatchingItem:
      type: object
      required: [videoId, video, progressSeconds, lastWatchedAt]
      properties:
        videoId:
          type: integer
        video:
          $ref: "#/components/schemas/Video"
        progressSeconds:
          type: integer
        lastWatchedAt:
          type: string
          format: date-time
    ContinueWatchingInput:
      type: object
      required: [progressSeconds]
      properties:
        progressSeconds:
          type: integer
          minimum: 0
    BookmarkInput:
      type: object
      required: [videoId]
      properties:
        videoId:
          type: integer
...
    Material:
      type: object
      required: [id, title, category, sourceUrl, pageCount]
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          $ref: "#/components/schemas/VideoCategory"
        sourceUrl:
          type: string
        pageCount:
          type: ["integer", "null"]
    MaterialBookmark:
      type: object
      required: [id, materialId, pageNumber, note, createdAt]
      properties:
        id:
          type: integer
        materialId:
          type: integer
        pageNumber:
          type: integer
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    MaterialBookmarkInput:
      type: object
      required: [pageNumber]
      properties:
        pageNumber:
          type: integer
          minimum: 1
        note:
          type: ["string", "null"]
    MaterialHighlight:
      type: object
      required: [id, materialId, pageNumber, text, color, createdAt]
      properties:
        id:
          type: integer
        materialId:
          type: integer
        pageNumber:
          type: integer
        text:
          type: string
        color:
          type: string
        createdAt:
          type: string
          format: date-time
    MaterialHighlightInput:
      type: object
      required: [pageNumber, text, color]
      properties:
        pageNumber:
          type: integer
          minimum: 1
        text:
          type: string
          minLength: 1
        color:
          type: string
...
    AiSessionMode:
      type: string
      enum: [explain, doubt, summarize, quiz]
    AiSession:
      type: object
      required: [id, title, mode, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        mode:
          $ref: "#/components/schemas/AiSessionMode"
        createdAt:
          type: string
          format: date-time
    AiSessionInput:
      type: object
      required: [title, mode]
      properties:
        title:
          type: string
          minLength: 1
        mode:
          $ref: "#/components/schemas/AiSessionMode"
    AiMessageRole:
      type: string
      enum: [user, assistant]
    AiMessage:
      type: object
      required: [id, sessionId, role, content, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        role:
          $ref: "#/components/schemas/AiMessageRole"
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    AiMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
...