# lib/api-zod/src/generated/types/materialHighlightInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * GROFO API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MaterialHighlightInput {
  /** @minimum 1 */
  pageNumber: number;
  /** @minLength 1 */
  text: string;
  color: string;
}